Extracted  ExtensionPoint and extension point store out into a standalone NPM package
Blue Ocean core will be able to create and "export" this (js-modules export). Plugins will "import" that shared instance, allowing them to register/add extension points etc etc.

Will need to enhance the render function to do the async loading of modules when it encounters extension points that are not yet registered.