This is a good example of how access control works in Jenkins
If the identity that's making the HTTP call has a specific permission (in this case Jenkins.ADMINISTER = god access) in the eyes of the currently installed AuthorizationStrategy, then this check returns true.

We could have defined a new permission here if we wanted, but Jenkins.ADMINISTER is sufficient. AuthorizationStrategy allows any user to have the god access on himself/herself, so anyone can
read any information about himself.

What this allow you to do, in addition, is that for example when CJP RBAC plugin is in place, there can be ops team lead "ben" account that has ADMINISTER permission on every User object, and those decision making logic is decoupled from the application logic.