Updated js-builder dep to 0.0.29 and removed setting of mvn build versions of node and npm
I fixed the issue that was causing js-builder to require a newer version of npm. The issue was that it was specifying a Browserify transformer by name Vs resolving locally and passing an instance. By-name resolution looks for the package in the top level of node_modules, causing it to bomb with older versions (that don't flatten the tree).