[JENKINS-37141] rework Docker build for development image - take 2 (#513)
* [JENKINS-37141] rework Docker build for development image - take 2

I just discovered hpi:assemble-dependencies, which is exactly what is
needed to bundle all plugins into a folder with can be used for the
Docker build. A dedicated module is also not needed at all (it was
required by hpi:bundle-plugins). It's a bit unfortunate to discover
this now, but better late than never..

Although matrix-auth and antisamy-markup-formatter are not needed to run
BlueOcean, it's good to have them around when the development image is
started with demo data.