JENKINS-38335# Slow search API response fix
numberOfRunningPipelines is very expensive call, pretty much results in to
iteraring over all builds in the pipeline. With Search API, it gets worse as we
recursively iterate over each job/pipeline and determine if its running or not.
There is no esay way in Jenkins to compute this value optimally.

There is no UI code that needs this field and also numberOfQueuedPipelines field
is not used. As part of this fix, I am removing these fields from Pipeline API response.