JENKINS-44624# Fix to avoid multiple open blueocean menu
When there are more than one implementation of BlueOceanActionUrl which
implements RootAction, there would (of course) be multiple menu items on
the left menu.

This fix changes the abstraction where BlueOceanUrlAction is a final
class and BlueOceanUrlObjectFactory is factory of BlueOceanUrlObject,
which defines properties to render 'Open Blue Ocean' left side menu.