[JENKINS-45703][JENKINS-45442] [JENKINS-45482] [JENKINS-45360] Several fixes around organizations (#1272)
* [JENKINS-45482] unit test showing incorrect behaviour.

If a custom blue ocean organisation is in use and it uses a root item
that is not Jenkins then modifying existing GitHub organistaions fails.

* [JENKINS-45442] unit test showing incorrect permission check.

When a custom organisation is in use pipeline creation checks
permissions on jenkins rather than the organistion root.

* Though shalt not set system properties in tests without restting them

* [JENKINS-45360] unit test for duplicate MBP with custom org.

Improve the test coverage and demonstrate the issue with MBP and custom
organizations.
shouldFailOnValidation4 is the test that fails in this case.

* Just increasing test coverage of custom orgs.

No issue was observed around this area but given I was working in the
area seemed logical to add it at the same time to prevent any
regreessions.

* [JENKINS-45422] the ACL check should be on the org folder.

The ACL check was checking the wrong ACL leading to a failure to create
pipelines when a custom organisation was in use.

* [JENKINS-45482] Use the correct Item.

When obraining the Item to see if it already exists the check used the
incorrect path if a custom Organisation was in use.

* [JENKINS-45360] check the organisation root not Jenkins for duplicates.

When we are checking for existing items we need to do it in the root of the organisation and not the root of Jenkins.

* Credentials uses an invalid org when no authentication in use

* [JENKINS-45703] - Check the organization from the url and the factory instead of jenkins root

* Adress comment.

* change version to steel

* [JENKINS-45703] Use better regex

* revert to normal snapshotting names