JENKINS-46156 Only attempt to stop a run once when the user clicks the stop link.
Previously, Blue Ocean would send the stop signal to a job multiple times until the job completed.
This did not allow jobs to handle aborts properly, for example to perform clean up steps if the job was aborted.

Now, Blue Ocean will send the stop signal once, and wait for the job to complete.