[JENKINS-46452] Don't reference null Jira session
[JENKINS-46452] reports null pointer exceptions which cause build failures if the Jira plugin is configured without credentials.

I use the Jira plugin to rewrite bug report URL's, but do not want it to submit updates to the bugs.  This change should allow my use model to work again.

My apologies that I have not written a test for the bug.  I'm OK if you reject this change in favor of a change which includes a test that shows the failure.