Add focusable="false" to StatusIndicator
All SVG need to have focusable="false" otherwise IE may give them focus even if they are not anymore in the DOM and your document.activeElement becomes null and TAB_INDEX starts at top level (either document or browser(!)).

Can be seen in any creation flow. When using the keybaord to navigate in the create flow if you tab from an input field you get nothing and then tab again and you get the next element in the focus.

Basically you need to shift-tab or tab twice to move between the control elements when a statusIndicator is in between since it will "steal" the focus due the the svg not having focusable="false".

Happens in IE only, Edge is correct as is Chrome.