[JENKINS-47146,JENKINS-45589] Remove expensive fallback path for expected build number and NPE (#1433)
* Remove expensive fallback path for expected build number and NPE
This changes fixes an issue I've seen come up:
    - When large numbers of triggers are run at once, the getBuilds call in getRun is extremely expensive.  It effectively ends up loading all the builds for a certain job.
    - This causes delays of sometimes several hours in scheduling jobs when load is high.
    - NPE when queued items can't be found in the queue.  This tends to happen more on startup.
The reason for removing this code is basically like this.  When a queue item that is blocked, waiting, etc. passes through here, the codepath appears to always return null.  This is becuase the queue item has not yet become a build and thus there is no entry in the job's build list.  The one I've found case where this does work is when the queue item leaves the queue (becomes a LeftItem).  At that point, it has been assigned a run number, and this code does work.  However, logically at that point this is no longer an expected build number.  It's an actual build number.  The regular build id logic in the blue ocean UI should take over at that point.

Now, it's entirely possible that I missed a crucial case here.  However, the getRun code is undoubtedly incorrect as it is extremely inefficient.

* Check for null job_run_queueId

* Check job event, if its running then simply return run id