[JENKINS-47158] Ensure that synthetic FlowNode is not saved.
This is just a bandaid - as discussed on JENKINS-47158, the usage of
synthetic `FlowNode`s for visualization-only purposes is not
ideal. But at a minimum, this prevents the serialization of the
synthetic `FlowNode`s, making JENKINS-45892 warnings go away, and not,
well, serializing a bunch of stuff we almost certainly shouldn't be.