Add a Retry to the Bitbucket e2e test (#1718)
Thanks guys. I think V2 of this effort will be to figure out why it's failing occaionally-but-not-always. I'd like to at least make the selector that saveBranch() is looking for. It's been on my list of ones to fix up for a while now, and since saveBranch() is used all over the place, probably a good time.