[JENKINS-51627] the code was just plain wrong and ugly :)
The code attempted to make sure that only one "Try BlueOcean" action was shown butno one looked at the reason why multiple where shown.
The reason that multiple actions where shown was a bug in core that affected Queue.Items and their subclasses.  Given that almost all Queue.Items have no page of their own there is little point in creating an Action for them (and this solves the fact that they get persisted and then shown for a AbstractBuild.

In the case they are already persisted we just replace the action with an invisibleAction such that the only rendered "Try Blue" will be the one created afresh from the TransientActionFactory