Improve Jenkinsfile error handling by removing unnecessary catch block
Previous usage of the echo step was incorrect because the argument
was not a string. It is generally better to throw errors instead of
logging them and using `currentBuild.result` to set the result so
that the stack trace of the error is displayed in the build log.

We use `catchError` in the `finally` block so that exceptions thrown
by that stage do not mask errors throw from the `try` block, which
are more likely to be the root cause of any problems.