[JENKINS-62767] Don't bundle git test deps in events plugin
The blueocean-events plugin has a test dependency on the git plugin.
The maven hpi plugin incorrectly includes the transitive dependencies
of test dependencies in the hpi file.

That caused the blueocean-events plugin to deliver a copy of JGit
4.5.4 and its dependencies when the git client plugin (owner of the
JGit library) is now delivering JGit 5.8.0.

Might also be able to resolve the issue by updating the maven hpi
plugin to a newer version or by updating the parent pom to a newer
version.  Since I can't compile the blueocean-plugin on my computer,
this seemed like the smallest possible change that would remove the
undesirable inclusions from the hpi file.  This technique can also be
tested by viewing the pull request build log to confirm that
blueocean-events no longer includes JGit.