exclude transative dep slf4j-api
so I think this is another bug in the HPI plugin in determining what to
bundle.

jenkins core depends on some slf4j libraries but not directly on
slf4j-api (which is wrong). so it is a transitive dep (3 level) and
scope provided.
slf4j-api is compile scope in jose4j so is a level 2 dep - as far as
maven dependencies are converned this wins and is promoted.
my view however is that hpi should be Jenkins aware and ignoring this.