Skip flaky multiBranchPipelineIndex test
The test failed on a plugin bill of materials run for 2 of the 4 tested
configurations.  The failure message was a

Caused by: java.nio.file.DirectoryNotEmptyException: blueocean-pipeline-api-impl/target/tmp/j h2197780848444070882/jobs/p

This is a rather heavy-handed way to avoid the intermittent test failure,
since it will stop running the test in all environments, all the time.

The other tests in this class that use the runAllTests flag have used
it since 2016.  I didn't find any reference that indicated the tests
are enabled by any operation in the repository itself.  They might
be executed on a CI system that defines the RUN_MULTIBRANCH_TESTS
environment variable.