[JENKINS-73775] Replace deprecated call to JGit setLastModified(long)
JGit 5 deprecated the DirCacheEntry.setLastModified(long) method and
recommends DirCacheEntry.setLastModified(Instant).  JGit 7.0.0 removes
DirCacheEntry.setLastModified(long).  This change will work with old
versions of JGit and JGit 7.0.0 because the replacement method is
available with older versions and the most recent release.

Javadoc for JGit 6.9.0 is available at
https://javadoc.io/doc/org.eclipse.jgit/org.eclipse.jgit/6.9.0.202403050737-r/org.eclipse.jgit/org/eclipse/jgit/dircache/DirCacheEntry.html

Without this change, users that try to run the Git portion of Blue Ocean
may see a method not found exception that will fail the operation.