Prepare for Java 25
Java 25 will release Sep 16, 2025 and the Jenkins project plans to
support it soon after its release.  This upodate to the most recent
parent pom successfully compiles and tests with Java 25 early access 26,
when combined with the removal of a few spotbugs suppressions that are
no longer necessary.

Switches the Jenkinsfile to test Java 25 and Java 21 instead of testing
Java 17 and Java 21.  Generated byte code is still Java 17, but I'm not
aware of any case where Java 17 exposed an issue that was not also seen
with Java 21.